Made default player constructor

+++ app/src/main/java/com/pipeecho/starpirates/WeaponClass.java
@@ -18,7 +18,7 @@ public class WeaponClass {
     String Description;
 
     int Level;
-    float LevelsPerUpgrade;
+    double LevelsPerUpgrade;
     int DamagePerClick;
     int ClicksPerClip;
     int ClickAmount;
